+++ src/main/java/uk/gov/justice/framework/command/client/MainApplication.java
@@ -1,5 +1,7 @@
 package uk.gov.justice.framework.command.client;
 
+import static uk.gov.justice.framework.command.client.ReturnCode.SUCCESS;
+
 import uk.gov.justice.framework.command.client.cdi.producers.OptionsFactory;
 import uk.gov.justice.framework.command.client.jmx.ListCommandsInvoker;
 import uk.gov.justice.framework.command.client.startup.CommandLineArgumentParser;
@@ -37,7 +39,7 @@ public class MainApplication {
     @Inject
     private ReturnCodeFactory returnCodeFactory;
 
-    public int run(final String[] args) {
+    public ReturnCode run(final String[] args) {
 
         final Optional<CommandLine> commandLineOptional = commandLineArgumentParser.parse(args);
 
@@ -60,6 +62,6 @@ public class MainApplication {
             formatter.printHelp("java -jar catchup-shuttering-manager.jar", optionsFactory.createOptions());
         }
 
-        return 0;
+        return SUCCESS;
     }
 }