+++ src/main/java/uk/gov/justice/framework/command/client/startup/CatchUpAndShutteringManager.java
@@ -3,6 +3,6 @@ package uk.gov.justice.framework.command.client.startup;
 public class CatchUpAndShutteringManager {
 
     public static void main(String... args) {
-        System.exit(new Bootstrapper().startContainerAndRun(args));
+        System.exit(new Bootstrapper().startContainerAndRun(args).getCode());
     }
 }